Reorganize & Finish CRUD of Projects Table.
