+++ base-senior-project-main/backend/src/handlers/ProjectCRUD/getAllProjects.js
@@ -1,28 +1,32 @@
 'use strict'
 
+/*Used for getting every project in the projects table
+URL/getAllProjects
+*404: no items in table*
+*200: good request/items returned*
+*500: error*
+*/
+
 const AWS = require('aws-sdk');
 const dynamoDb = new AWS.DynamoDB.DocumentClient();
 module.exports.handler = async (event) => {
     const params = {
         TableName: process.env.PROJECTS_TABLE,
-        Key: {
-            ProjectID: parseInt(event.pathParameters.ProjectID)
-        }
     };
 
     try {
-        const data = await dynamoDb.get(params).promise();  // Query DynamoDB using the projectID
+        const data = await dynamoDb.scan(params).promise();  // Query DynamoDB for all projects
         
-        if (!data.Item) {
+        if (!data.Items || data.Items.length === 0) {
             return {
                 statusCode: 404,
-                body: JSON.stringify({ message: 'Project not found' }),
+                body: JSON.stringify({ message: 'Projects not found' }),
             };
         }
 
         return {
             statusCode: 200,
-            body: JSON.stringify(data.Item),  // Return the project data as JSON
+            body: JSON.stringify(data.Items),  // Return the projects data as JSON
         };
     } catch (err) {
         return {
